move away RLP internal tables from table.rs
